Updated Entity

+++ PharmaDAL/Entity/PurchaseSaleBookLineItem.cs
@@ -23,8 +23,8 @@ namespace PharmaDAL.Entity
         public string ItemCode { get; set; }
         public string Batch { get; set; }
         public string BatchNew { get; set; }
-        public int Quantity { get; set; }
-        public Nullable<int> FreeQuantity { get; set; }
+        public double Quantity { get; set; }
+        public Nullable<double> FreeQuantity { get; set; }
         public double PurchaseSaleRate { get; set; }
         public Nullable<double> EffecivePurchaseSaleRate { get; set; }
         public string PurchaseSaleTypeCode { get; set; }
